+++ pages/globalSearchPage.js
